opt/fazai/worker_python/requirements.txt
llama-cpp-python>=0.3.0
pydantic>=2.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart>=0.0.6
httpx>=0.25.0
qdrant-client>=1.7.0
sentence-transformers>=2.2.0
openai>=1.0.0
aiohttp>=3.9.0
prometheus-client>=0.19.0
psutil>=5.9.0
